requirements.txt
Django==5.0.2
django-allauth==0.57.0
django-crispy-forms==2.0
crispy-tailwind==0.5.0
Pillow==10.2.0
python-decouple==3.8
channels==4.0.0
channels-redis==4.1.0
redis==5.0.0
django-redis==5.3.0
mysqlclient==2.2.4
